requirements.txt
# =========================
# CORE
# =========================
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2

# =========================
# STABLE DIFFUSION
# =========================
diffusers==0.25.1
transformers==4.36.2
huggingface_hub==0.20.3
accelerate==0.25.0
safetensors==0.4.1
xformers==0.0.23.post1; platform_system != "Windows"

# =========================
# CONTROLNET / IMAGE
# =========================
opencv-python-headless
pillow
numpy

# =========================
# FACE ANALYSIS
# =========================
insightface==0.7.3
onnxruntime-gpu==1.16.3

# =========================
# API
# =========================
fastapi
uvicorn[standard]
pydantic

# =========================
# UTILS
# =========================
tqdm
scipy
